+++ autorest/preparer_test.go
@@ -760,7 +760,39 @@ func TestModifyingExistingRequest(t *testing.T) {
 	if err != nil {
 		t.Fatalf("autorest: Preparing an existing request returned an error (%v)", err)
 	}
-	if r.URL.String() != "https:/search?q=golang" && r.URL.Host != "bing.com" {
-		t.Fatalf("autorest: Preparing an existing request failed (%s)", r.URL)
+	if r.URL.Host != "bing.com" {
+		t.Fatalf("autorest: Preparing an existing request failed when setting the host (%s)", r.URL)
+	}
+
+	if r.URL.Path != "/search" {
+		t.Fatalf("autorest: Preparing an existing request failed when setting the path (%s)", r.URL.Path)
+	}
+
+	if r.URL.RawQuery != "q=golang" {
+		t.Fatalf("autorest: Preparing an existing request failed when setting the query parameters (%s)", r.URL.RawQuery)
+	}
+}
+
+func TestModifyingRequestWithExistingQueryParameters(t *testing.T) {
+	r, err := Prepare(
+		mocks.NewRequestForURL("https://bing.com"),
+		WithPath("search"),
+		WithQueryParameters(map[string]interface{}{"q": "golang the best"}),
+		WithQueryParameters(map[string]interface{}{"pq": "golang+encoded"}),
+	)
+	if err != nil {
+		t.Fatalf("autorest: Preparing an existing request returned an error (%v)", err)
+	}
+
+	if r.URL.Host != "bing.com" {
+		t.Fatalf("autorest: Preparing an existing request failed when setting the host (%s)", r.URL)
+	}
+
+	if r.URL.Path != "/search" {
+		t.Fatalf("autorest: Preparing an existing request failed when setting the path (%s)", r.URL.Path)
+	}
+
+	if r.URL.RawQuery != "pq=golang+encoded&q=golang+the+best" {
+		t.Fatalf("autorest: Preparing an existing request failed when setting the query parameters (%s)", r.URL.RawQuery)
 	}
 }